requirements.txt
argparse
yt-dlp
selenium
ultralytics
torch
torchvision
Pillow
numpy
lmdb
tqdm
scikit-learn
opencv-python
tensorboard
